use patch decorator in various places
